aaaaaa

+++ RelacionamentoHeranca/Models/Perecivel.cs
@@ -1,9 +1,10 @@
 ﻿namespace RelacionamentoHeranca.Models
+
 {
     public class Perecivel : Produto
     {
         public DateTime DataValidade { get; set; }
         public string? Sabor { get; set; }
-        public float Peso { get; set; }
+        public double Peso { get; set; }
     }
 }
